Removed collection name from the generated typdefs.
The collection name is an unreliable witness. Some structures
comprise more than one logical collection, so instead of
choosing one of the collection names to name include in the
structure (typedef) name it is cleaner to leave the collection
out. Instead collection names are inserted as comments before
each group of fields enclosed by that collection.