Added usage 0x56 (Relative Scan Time) - Microsoft extension
Reports relative time in units of 100 microseconds.

It represents the delta from the first frame that was reported after a device
starts reporting data subsequent to a period of inactivity. The first scan
time received is treated as a base time for subsequent reported times. The
deltas between reported scan times should reflect the scanning frequency of
the digitizer. It is important to note that unlike other usages, the host
does not allow any flexibility for the unit for the scan time usage. The
value is expected to roll over, as only 1 byte is allocated to the counter.
The scan time value should be the same for all contacts within a frame.
This requirement also applies to devices that report data using the hybrid mode