Reworked UNIT and UNIT_EXPONENT handling.
The formula used to convert logical units into physical units is
now displayed as a comment next to each field in the C structure
output. Instead of just displaying the formula verbatim, an attempt
is made to reduce it to a simpler form. For example, instead of:
  physical = (logical - 0) x 256000/256 + 0
the formula is reduced to:
  physical = logical x 1000
It's not perfect but better than the former.