Issue a warning at the end of each collection if physical units are still in effect.
This is a common mistake made when creating HID Report Descriptors. The author
should either surround the usage of physical units with PUSH and POP, or should
explicitly reset PHYSICAL_MINIMUM, PHYSICAL_MAXIMUM, UNIT and UNIT_EXPONENT to
zero once the physical units are no longer needed for an INPUT, OUTPUT or FEATURE report.
Failing to do that will cause the physical units to be applied to ALL subsequent
USAGES (for example, a count value may have "metres per second" applied to it).
That is because physical units are GLOBAL values that do not vary unless explicitly
changed.