Added rudimentary #define support.
This only works if the #define statements are in the file being parsed
(this software is not a C compiler). If any #define statements are for
decimal or hexadecimal numbers, and the identifiers are found while
searching for hex strings in a report descriptor, then those
identifiers will be replaced by the corresponding value on the #define
statements.