Properly handle the case where USAGE_MAXIMUM is specified before USAGE_MINIMUM.
USAGE_MINIMUM and USAGE_MAXIMUM specify a range of usages - which is more
efficient than specifying individual usages if you want to specify more than 2
adjacent usages.

Until now the USAGE_MINIMUM had to be specified first because the range was
actioned when the USAGE_MAXIMUM item was encountered. Now the range is actioned
when it is determined that both minimum and maximum have been specified.